+++ src/_pytask/hookspecs.py
@@ -18,7 +18,6 @@ if TYPE_CHECKING:
     import click
     from pluggy import PluginManager
 
-    from _pytask.logging_utils import TaskExecutionStatus
     from _pytask.models import NodeInfo
     from _pytask.node_protocols import PNode
     from _pytask.node_protocols import PProvisionalNode
@@ -256,9 +255,7 @@ def pytask_execute_task_protocol(session: Session, task: PTask) -> ExecutionRepo
 
 
 @hookspec(firstresult=True)
-def pytask_execute_task_log_start(
-    session: Session, task: PTask, status: TaskExecutionStatus
-) -> None:
+def pytask_execute_task_log_start(session: Session, task: PTask) -> None:
     """Start logging of task execution.
 
     This hook can be used to provide more verbose output during the execution.