+++ src/_pytask/live.py
@@ -20,12 +20,12 @@ from rich.text import Text
 
 from _pytask.console import console
 from _pytask.console import format_task_name
+from _pytask.logging_utils import TaskExecutionStatus
 from _pytask.outcomes import CollectionOutcome
 from _pytask.outcomes import TaskOutcome
 from _pytask.pluginmanager import hookimpl
 
 if TYPE_CHECKING:
-    from _pytask.logging_utils import TaskExecutionStatus
     from _pytask.node_protocols import PTask
     from _pytask.reports import CollectionReport
     from _pytask.reports import ExecutionReport
@@ -151,6 +151,7 @@ class LiveExecution:
     n_entries_in_table: int
     verbose: int
     editor_url_scheme: str
+    initial_status: TaskExecutionStatus = TaskExecutionStatus.RUNNING
     sort_final_table: bool = False
     n_tasks: int | str = "x"
     _reports: list[_ReportEntry] = field(factory=list)
@@ -170,11 +171,9 @@ class LiveExecution:
         return result
 
     @hookimpl(tryfirst=True)
-    def pytask_execute_task_log_start(
-        self, task: PTask, status: TaskExecutionStatus
-    ) -> bool:
+    def pytask_execute_task_log_start(self, task: PTask) -> bool:
         """Mark a new task as running."""
-        self.add_task(new_running_task=task, status=status)
+        self.add_task(new_running_task=task, status=self.initial_status)
         return True
 
     @hookimpl